fix(MangaTaro-QToon): move domain constants to shared models
