+++ src/MangaTaro/implementations/manga/main.ts
@@ -1,6 +1,6 @@
 import type { Request, SourceManga } from "@paperback/types";
 import { URL } from "@paperback/types";
-import { DOMAIN } from "../../main";
+import { DOMAIN } from "../shared/models";
 import { fetchText } from "../../services/network";
 import { parseMangaId } from "../shared/utils";
 import { parseMangaDetails } from "./parsers";
