+++ src/MangaTaro/implementations/search-results/main.ts
@@ -7,7 +7,7 @@ import type {
   SortingOption,
 } from "@paperback/types";
 import { ContentRating, URL } from "@paperback/types";
-import { DOMAIN } from "../../main";
+import { DOMAIN } from "../shared/models";
 import { fetchJSON } from "../../services/network";
 import type { MangaTaroLoadItem, MangaTaroLoadRequest } from "../shared/models";
 import { formatMangaId, isNovel, slugFromUrl } from "../shared/utils";
