+++ src/MangaTaro/services/network.ts
@@ -1,6 +1,6 @@
 import type { Request, Response } from "@paperback/types";
 import { CloudflareError, PaperbackInterceptor } from "@paperback/types";
-import { DOMAIN } from "../main";
+import { DOMAIN } from "../implementations/shared/models";
 
 export class MangaTaroInterceptor extends PaperbackInterceptor {
   async interceptRequest(request: Request): Promise<Request> {
