+++ src/QToon/implementations/discover-section/main.ts
@@ -1,6 +1,6 @@
 import type { DiscoverSection, DiscoverSectionItem, PagedResults, Request } from "@paperback/types";
 import { URL } from "@paperback/types";
-import { DOMAIN_API } from "../../main";
+import { DOMAIN_API } from "../shared/models";
 import { fetchEncryptedJSON } from "../../services/network";
 import type {
   DiscoverMetadata,
