+++ src/QToon/implementations/manga/parsers.ts
@@ -1,6 +1,6 @@
 import type { SourceManga } from "@paperback/types";
 import { ContentRating } from "@paperback/types";
-import { DOMAIN } from "../../main";
+import { DOMAIN } from "../shared/models";
 import type { QToonComic } from "../shared/models";
 import { comicId } from "../shared/utils";
